Updated WILC driver to RTP version 15.0 * Bug Fixes * WILC3000 WFA Certification *

+++ drivers/net/wireless/atmel/wilc1000/wilc_spi.c
@@ -32,7 +32,7 @@ static struct wilc_spi g_spi;
 
 static int spi_read(uint32_t, uint8_t *, uint32_t);
 static int spi_write(uint32_t, uint8_t *, uint32_t);
-uint8_t spi_reset(void);
+int spi_reset(void);
 
 /*
  * Crc7
@@ -1156,7 +1156,7 @@ static void spi_init_pkt_sz(uint32_t* reg)
 	}
 }
 
-uint8_t spi_reset(void)
+int spi_reset(void)
 {
 	int result = N_OK;
